+++ src/main/java/com/opus/opus/global/security/SecurityConfig.java
@@ -16,9 +16,6 @@ import org.springframework.security.config.annotation.web.configuration.EnableWe
 import org.springframework.security.config.annotation.web.configurers.CsrfConfigurer;
 import org.springframework.security.config.annotation.web.configurers.FormLoginConfigurer;
 import org.springframework.security.config.annotation.web.configurers.HttpBasicConfigurer;
-import org.springframework.beans.factory.annotation.Value;
-import org.springframework.security.crypto.encrypt.Encryptors;
-import org.springframework.security.crypto.encrypt.TextEncryptor;
 import org.springframework.security.crypto.factory.PasswordEncoderFactories;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.security.oauth2.client.registration.ClientRegistrationRepository;
@@ -41,19 +38,9 @@ public class SecurityConfig {
     private final GoogleOAuth2LoginFailureHandler googleOAuth2LoginFailureHandler;
     private final ClientRegistrationRepository clientRegistrationRepository;
 
-    @Value("${aes.secret-key}")
-    private String aesSecretKey;
-    @Value("${aes.salt}")
-    private String aesSalt;
-
     @Bean
     public PasswordEncoder passwordEncoder() {
         return PasswordEncoderFactories.createDelegatingPasswordEncoder();
-    }
-
-    @Bean
-    public TextEncryptor textEncryptor() {
-        return Encryptors.delux(aesSecretKey, aesSalt);
     }
 
     @Bean
